Move read/write device state to utils
Since multiple programs use this, it should be in utils.
