+++ src/touch_sensor/touch_sensor.py
@@ -8,9 +8,9 @@ import gpiod
 
 from src.cloud_db.cloud_db import CloudDBClient
 from config.touch_sensor_config import *
-from src.utils.paths import (TOUCH_SENSOR_TO_SCREEN_PIPE, PROJECT_DIR)
+from src.utils.paths import (DEVICE_STATE_PATH, TOUCH_SENSOR_TO_SCREEN_PIPE, PROJECT_DIR)
 from src.utils.touch_sensor_utils import *
-from src.utils.utils import get_device_id, TapStatus
+from src.utils.utils import read_device_state, get_device_id, TapStatus
 
 logging.config.fileConfig(f"{PROJECT_DIR}/config/logging.conf", disable_existing_loggers=False)
 logger = logging.getLogger(os.path.basename(__file__).split('.')[0])
