+++ src/utils/screen_utils.py
@@ -125,66 +125,6 @@ def is_at_least_next_day(from_timestamp: float, to_timestamp: float) -> bool:
     # Check if to_timestamp is at least the next day from from_timestamp
     return to_day_start > from_day_start
 
-# TODO: Make sure that this function handles concurrent access
-def read_device_state(path_to_device_state: str, verbose: bool = True) -> Dict[str, Any]:
-    """Read the device state from a JSON file.
-
-    Args:
-        path_to_device_state (str): Path to the JSON file containing the device state.
-
-    Returns:
-        Dict[str, Any]: A dictionary representing the device state.
-    
-    Raises:
-        FileNotFoundError: If the specified JSON file path does not exist.
-        JSONDecodeError: If there is an error decoding the JSON data from the file.
-    """
-    if verbose:
-        logger.info('Reading device state')
-    try:
-        with open(path_to_device_state, 'r') as file:
-            portalocker.lock(file, portalocker.LOCK_SH)
-            data = json.load(file)
-            portalocker.unlock(file)
-            return data
-    except FileNotFoundError:
-        logger.error('Device state file not found')
-        raise FileNotFoundError
-    except json.JSONDecodeError:
-        logger.error('Unable to parse device state: Invalid JSON format')
-        raise json.JSONDecodeError
-    except portalocker.exceptions.BaseLockException as e:
-        print(f"Failed to lock the file: {e}")
-
-# TODO: Make sure that this function handles concurrent access
-def write_device_state(device_state: Dict[str, Any], path_to_device_state: str, verbose: bool = True) -> None:
-    """Write the updated device state to a JSON file.
-
-    Args:
-        device_state (Dict[str, Any]): The device state to write.
-        path_to_device_state (str): Path to the JSON file where the device state will be saved.
-    
-    Raises:
-        FileNotFoundError: If the specified JSON file path does not exist.
-        IOError: If there is an error writing to the file.
-    """
-    if verbose:
-        logger.info('Writing device state')
-    device_state['saved_timestamp'] = time.time()
-    try:
-        with open(path_to_device_state, 'w') as file:
-            portalocker.lock(file, portalocker.LOCK_EX)
-            json.dump(device_state, file, indent=4)
-            portalocker.unlock(file)
-    except FileNotFoundError:
-        logger.error('Device state file not found')
-        raise FileNotFoundError  # TODO: Determine error message format
-    except IOError as e:
-        logger.error(f"An error occurred while writing to device state: {e}")
-        raise IOError  # TODO: Determine error message format
-    except portalocker.exceptions.BaseLockException as e:
-        print(f"Failed to lock the file: {e}")
-
 def read_pipe(path_to_pipe: str) -> Dict[str, Any]:
     """Read and parse JSON data from a named pipe.
